+++ CHANGELOG.md
@@ -6,10 +6,15 @@
 
 - Function for getting lanelet orientation closest to a given position
 - Function for getting most likely lanelet given an obstacle state
+- Function for erasing lanelet network from scenario
+- Function for replacing lanelet network of a scenario with new one
 
 ### Changed
 
 - Move tests, tutorial, and documentation folder to root directory
+
+### Removed
+- setter method for lanelet network in scenario class
 
 ### Fixed